% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPRDcolours.R
\name{bprd_pal}
\alias{bprd_pal}
\alias{view_bprd_pals}
\title{BPRD color palettes}
\usage{
bprd_pal(palette, reverse = FALSE)

view_bprd_pals(palette)
}
\arguments{
\item{palette}{character, name of predefined palette. If missing, return all available palette names.}

\item{reverse}{logical, reverse color order.}
}
\value{
character vector of hex colors or palette names
}
\description{
Return a predefined BPRD color palette from the \code{BPRDpals} dataset.
}
\details{
Many of the palettes are qualitative, and not necessarily evenly spaced in terms of hue, saturation or brightness.
This is because many palettes come from logos, symbols, insignia and other simple representations.
However, several palettes have specifically been constructed as sequential or divergent if it made sense to do so based on the dominant colors present.
}
\examples{
bprd_pal("hellboy")
# leave palette blank to list available names:
bprd_pal()

# to view all palettes
view_bprd_pals()
}
