% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPRDcolours.R, R/data.R
\docType{data}
\name{BPRDpals}
\alias{BPRDpals}
\alias{hellboy}
\alias{abe}
\alias{liz}
\alias{roger}
\alias{krauss}
\title{BPRD color palettes.}
\format{
A named list.

An object of class \code{data.frame} with 4 rows and 2 columns.

An object of class \code{data.frame} with 4 rows and 2 columns.

An object of class \code{data.frame} with 4 rows and 2 columns.

An object of class \code{data.frame} with 4 rows and 2 columns.

An object of class \code{data.frame} with 1 rows and 2 columns.

An object of class \code{list} of length 6.
}
\usage{
BPRDpals

hellboy

abe

liz

roger

krauss

BPRDpals
}
\description{
A named list of X BPRD color palettes.
}
\keyword{datasets}
