% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPRDcolours.R
\name{scale_bprd}
\alias{scale_bprd}
\alias{scale_colour_bprd}
\alias{scale_color_bprd}
\alias{scale_fill_bprd}
\title{Color and fill scale functions for BPRD palettes}
\usage{
scale_colour_bprd(palette = "abe", discrete = TRUE, reverse = FALSE, ...)

scale_color_bprd(...)

scale_fill_bprd(palette = "abe", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{character, name of BPRD palette. See \code{bprd_pal()} for list of palette names.}

\item{discrete}{logical, discrete or continuous palette.}

\item{reverse}{logical, reverse color order.}

\item{...}{additional arguments passed to \code{ggplot2::discrete_scale} or \code{ggplot2::scale_*_gradientn}, for discrete or continuous palettes, respectively.}
}
\description{
Scale functions used with ggplot2.
}
\details{
Most palettes should be used as qualitative palettes. See \code{BPRDpals} to see how many colors are in each predefined palette.
Use \code{view_bprd_pals()} to plot all palettes to see which may work best for your purposes.
}
\examples{
library(ggplot2)
d <- diamonds[diamonds$cut >= "Very Good", ]
ggplot(d, aes(carat, stat(count), fill = cut)) +
  geom_density(position = "fill") +
  scale_fill_bprd("abe")
}
